feat: ClientWOrdFilter周りの実装を切り出し
